Update subscription.py to allow following calendar months w/ yearly sub
Right now there is a check for the existence of a subscription end date when "follow calendar months" is enabled. It complains about the absence of a subscription end date when I try to create a subscription on a yearly billing cycle This seems unintended (though I can't tell from the documentation why this check exists at all), so I've proposed a fix that otherwise leaves the logic in place.

In this second commit I added logic for the invoice date generation.